docs: :memo: meta added
