+++ packages/portal/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@bot-whatsapp/portal",
-    "version": "0.0.9-alpha.0",
+    "version": "v0.1.11",
     "description": "Portal WEB para escanear QR",
     "main": "./lib/portal.http.cjs",
     "scripts": {
